requirements.txt
# 核心依赖
requests>=2.31.0

chromadb==0.5.20
tavily-python==0.3.6
pypdf==3.17.0
python-docx==1.1.0
sentence-transformers==5.2.0

flask==3.1.2
gradio==6.1.0
openai==1.51.0
sqlite3
python-dotenv==1.2.1
uvicorn==0.38.0
fastapi==0.115.2
aiohttp==3.9.1
langchain==0.3.27
langchain-chroma==0.2.6
langchain-community==0.3.27
langchain-core==0.3.81
langchain-openai==0.3.33
langchain-tavily==0.2.11
langchain-text-splitters==0.3.11

wave
hmac
hashlib

# 音频处理
ffmpeg-python>=0.2.0
struct